base/serializers.py
from rest_framework import serializers
from .models import Supplier

class SupplierSerializer(serializers.ModelSerializer):
    class Meta:
        model = Supplier
        fields = ['id', 'supplier', 'address', 'country', 'workers', 'sector', 'year', 'source_business']