requirements.txt
django
pandas
PyPDF2
djangorestframework>=3.13
psycopg2-binary>=2.9
django-cors-headers